+++ .vscode-test.mjs
@@ -3,3 +3,5 @@ import { defineConfig } from '@vscode/test-cli';
 export default defineConfig({
 	files: 'out/test/**/*.test.js',
 });
+
+